requirements.txt
click==8.1.7
pyyaml==6.0.1
netifaces==0.11.0